+++ simulate/bed.go
@@ -5,8 +5,8 @@ import (
 	"github.com/vertgenlab/gonomics/numbers"
 )
 
-func GoSimulateBed(searchSpace []*bed.Bed, regionCount int, regionLength int64) <-chan bed.Bed {
-	var Length, tmp, chromWindows int64
+func GoSimulateBed(searchSpace []*bed.Bed, regionCount int, regionLength int) <-chan bed.Bed {
+	var Length, tmp, chromWindows int
 	var totalWindows int
 	c := make(chan bed.Bed, 1000)
 
@@ -15,7 +15,7 @@ func GoSimulateBed(searchSpace []*bed.Bed, regionCount int, regionLength int64) 
 		Length = searchSpace[i].ChromEnd - searchSpace[i].ChromStart
 
 		if Length >= regionLength {
-			totalWindows = totalWindows + int(Length-regionLength+1)
+			totalWindows = totalWindows + (Length - regionLength + 1)
 		}
 	}
 
@@ -23,7 +23,7 @@ func GoSimulateBed(searchSpace []*bed.Bed, regionCount int, regionLength int64) 
 	//this function generates new bed regions and sends them to a channel.
 	go func() {
 		for i := 0; i < regionCount; i++ {
-			tmp = int64(numbers.RandIntInRange(0, totalWindows))
+			tmp = numbers.RandIntInRange(0, totalWindows)
 			// DEBUG fmt.Printf("Random number is %d\n", tmp)
 			for j := 0; j < len(searchSpace); j++ {
 				// DEBUG fmt.Printf("ChromEnd:%v ChromStart:%v \n",noGap[j].ChromEnd,noGap[j].ChromStart)
